refactoring classes to injection package plus symplifying integration
