+++ webservices/server-integration/src/main/java/org/jboss/as/webservices/dmr/WSDeploymentActivator.java
@@ -25,10 +25,7 @@ package org.jboss.as.webservices.dmr;
 import org.jboss.as.ee.component.deployers.EEResourceReferenceProcessorRegistry;
 import org.jboss.as.server.DeploymentProcessorTarget;
 import org.jboss.as.server.deployment.Phase;
-import org.jboss.as.webservices.component.WSHandlerChainAnnotationProcessor;
 import org.jboss.as.webservices.deployers.AspectDeploymentProcessor;
-import org.jboss.as.webservices.deployers.JaxwsEndpointComponentDescriptionFactory;
-import org.jboss.as.webservices.deployers.JaxwsHandlerComponentDescriptionFactory;
 import org.jboss.as.webservices.deployers.WSDependenciesProcessor;
 import org.jboss.as.webservices.deployers.WSDescriptorDeploymentProcessor;
 import org.jboss.as.webservices.deployers.WSEJBIntegrationProcessor;
@@ -36,6 +33,9 @@ import org.jboss.as.webservices.deployers.WSJMSIntegrationProcessor;
 import org.jboss.as.webservices.deployers.WSModelDeploymentProcessor;
 import org.jboss.as.webservices.deployers.WebServiceContextResourceProcessor;
 import org.jboss.as.webservices.deployers.deployment.DeploymentAspectsProvider;
+import org.jboss.as.webservices.injection.JaxwsEndpointComponentDescriptionFactory;
+import org.jboss.as.webservices.injection.JaxwsHandlerComponentDescriptionFactory;
+import org.jboss.as.webservices.injection.WSHandlerChainAnnotationProcessor;
 import org.jboss.as.webservices.webserviceref.WSRefAnnotationProcessor;
 import org.jboss.as.webservices.webserviceref.WSRefDDProcessor;
 import org.jboss.logging.Logger;
