+++ webservices/server-integration/src/main/java/org/jboss/as/webservices/injection/WSEndpointHandlersMapping.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.jboss.as.webservices.component;
+package org.jboss.as.webservices.injection;
 
 import java.util.Collections;
 import java.util.HashMap;
