+++ webservices/server-integration/src/main/resources/META-INF/stack-agnostic-deployment-aspects.xml
@@ -83,7 +83,7 @@
     <property name="provides" class="java.lang.String">EndpointRecordProcessors</property>
   </deploymentAspect>
 
-  <deploymentAspect class="org.jboss.as.webservices.deployers.InjectionDeploymentAspect">
+  <deploymentAspect class="org.jboss.as.webservices.injection.InjectionDeploymentAspect">
     <property name="requires" class="java.lang.String">StackInstanceProvider</property>
     <property name="provides" class="java.lang.String">InjectionAwareInstanceProvider</property>
   </deploymentAspect>
